+++ packages/helic/lib/Helic/Data/YankConfig.hs
@@ -1,11 +1,12 @@
 {-# options_haddock prune #-}
--- |YankConfig Data Type, Internal
 
+-- | YankConfig Data Type, Internal
 module Helic.Data.YankConfig where
 
 data YankConfig =
   YankConfig {
-    agent :: Maybe Text
+    agent :: Maybe Text,
+    text :: Maybe Text
   }
   deriving stock (Eq, Show, Generic)
   deriving anyclass (Default)